* kpropd.c (authorized_principal): Cast argument to ispace() to int
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13376 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/slave/kpropd.c
@@ -676,11 +676,11 @@ authorized_principal(context, p, auth_etype)
 
 	    /* if the next character is not whitespace or nul, then
 	       the match is only partial.  continue on to new lines. */
-	    if (*ptr && !isspace(*ptr))
+	    if (*ptr && !isspace((int) *ptr))
 		continue;
 
 	    /* otherwise, skip trailing whitespace */
-	    for (; *ptr && isspace(*ptr); ptr++) ;
+	    for (; *ptr && isspace((int) *ptr); ptr++) ;
 
 	    /* now, look for an etype string. if there isn't one,
 	       return true.  if there is an invalid string, continue.
